fix(security): address HIGH/MEDIUM security findings
HIGH: Fix API key regex to match Anthropic format (sk-ant-api03-...)
- Updated regex from [a-zA-Z0-9] to [a-zA-Z0-9._\-] to match dashes
- Applied same fix to key- pattern

MEDIUM: Sanitize error messages in session.py
- Moved sanitize_error_message() to base.py (shared module)
- Import and use in session.py for task_logger and error_info
- Prevents sensitive data in error logs

LOW: Remove redundant stopFn in agent-manager.ts
- restartTask() already calls killTask() internally
- Removed double-kill during profile swaps

LOW: Use asyncio.get_running_loop() instead of get_event_loop()
- Future-proofing for Python 3.10+ deprecation
